fix logic of methods CheckValidParsing, CalculateNumbers and Run
