+++ src/Practice2_Basics/Program.cs
@@ -1,16 +1,10 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Practice2_Basics
+﻿namespace Practice2_Basics
 {
     internal class Program
     {
         static void Main(string[] args)
         {
-            Conditions_Task5 conditions_Task5 = new Conditions_Task5();
+            ConditionsTask5 conditions_Task5 = new ConditionsTask5();
             conditions_Task5.Run();
         }
     }
